class circle created
